Remove branch-ours alias
Unfortunately, this cannot isolate merged commits as intended. It is only
possible when merged commits does not change features code (which is not usually
the case).  Notice that if no feature branch drop has been made to the
base branch, it is trivial to isolate the changeset with a 'git merge --squash'
into the base.  Harder but safer (generally applicable), it is to analyse each
(linear) sequence of commits between merges.  Merges can be listed with 'git log
--merges --right-only BASE_BRANCH...' (with feature branch checked out). Then,
each sequence can be analysed with: 'git difftool -d PREV_MERGE..NEXT_MERGE^'
(with range 'LAST_MERGE..' for the latest).